nexus-breach/backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
websockets>=13.0
python-dotenv>=1.0.1
httpx>=0.27.0
pydantic>=2.10.0
